+++ challenge_04/data/products.json
@@ -19,11 +19,11 @@
   },
   {
     "id": 3,
-    "title": "Iphone 12 Pro Max",
+    "title": "Iphone 13 Pro Max",
     "description": "High quality",
     "price": 4500,
     "thumbnail": "",
     "code": "000A4",
-    "stock": 28
+    "stock": 18
   }
 ]
